requirements.txt
requests
pytest
pyhamcrest